feat(db): release error handling, db lock type, removed read-before-write (#2337)
0) With this change, the endpoint to create new releases (/release and
/api/release) will now return an error (status 400) when a manifest for
a non-existing environment was supplied.

1) Replaced in-memory locks with db locks

Note that this feature is considered experimental.

With this change, we have three options in the helm chart
("cd.lockType") to lock:
* "go": with an in-memory (golang) lock (default)
* "db": with a postgres advisory lock
* "none": without any locks

The first option was already implemented, the other two options are new.
The postgres locks are an improvement over the go-locks, because they
are not limited to the current pod/process. Apart from that they work
the same.

Both lock types ensure that only one destructive action (undeploy,
delete env from app, delete env) is running at a time.


2) Removed "Read-before-Writes"

There were 2 points where the cd-service read from the DB before
writing:
When adding an app to an environment, and when removing an app from an
environment.
Both have been simplified to one sql statement.


Ref: SRX-8JRR7Q