fix: Add openapi spec for future use (#2359)
This adds an openapi.yaml to define the REST API of the frontend
service.

This was done by listening on a new port (5555). On this new port we
support 3 operations for now:
* GET /openapi.yaml: returns the yaml to be used within swagger-ui
* OPTIONS /openapi.yaml: required to allow browser to ignore the CORS
rules
* GET /api/commit-deployments/{commitHash}: this works the same as on
the old port.

Additionally to "pkg/api" there is now also generated code in
"pkg/publicapi".
To generate both at once you can run "make -C pkg gen".

Ref: SRX-WLGQSK