fix: always run all 0.0.0 migrations (#2384)
This fixes a bug where the initial db migrations did not run completely.

This bug affected anyone who migrated from git to database with a
version "v" of kuberpult with:
13.2.0 <= v < 13.16.1

Another way to check if you're affected by this bug is to run this psql
script:
```
		SELECT created, appName, metadata, manifests, releaseVersion, environments
		FROM releases
		WHERE COALESCE(environments, '') = '' AND COALESCE(manifests, '') != '';
```
If this returns no rows you are fine.

For those affected by this bug, run the migrations again with the next
kuberpult release (via helm parameter "db.checkCustomMigrations: true").

Ref: SRX-A7COB9