fix: disable github concurrency limitations (#2407)
This changes the github workflow to not run only one action per PR at a
time.
Reasoning is that the cancellation via "cancel-in-progress: true" does
not work (it does not seem to cancel anything), so this just slows us
down. Note that also clicking on "cancel" in the workflow also does not
work properly at the moment.

Ref: SRX-JEU54H