fix: change trains to load manifests only once in the beginning (#2405)
For "experimentalParallelismOneTransaction: true", this changes the
release train, so that loading all manifests of all releases only
happens once in the main go routine, and not in each spawned go routine.

Also introduces new spans to visualize trains better in datadog,
especially the part where the main go routine waits for all prognoses to
be finished.

Ref: SRX-JEU54H