fix: coverage gate for all services (#2435)
Added a coverage gate for each service (incl pkg and cli):

Each service has a threshold for code coverage in percent.
If a service is below this threshold, then the build fails.
If a service is way over this threshold (+5%), the build fails - in this
case the threshold must be increased.

Ref: SRX-UGK40K