fix(manifest-export): CVE-2025-58050 by removing pcre2 (#2641)
Previously, pcre2 was actually used by 2 different dependencies:
git - which was just installed in order to clone libgit2
libgit2 - which did use pcre2, but this is not a hard requirement.

With this change, the installation of "git" is removed after it's not
needed anymore.
And libgit2 is built without pcre2, so that the resulting image of the
manifest-export does not contain pcre2 anymore.

Ref: SRX-RL35GY