feat: new undeploy UI (#2666)
Refactor the environment deletion flow by replacing an inline selection
dialog with a dedicated, stateful component (EnvDelDialog). This new
component manages the selection of environments and directly
adds/removes the relevant deleteEnvFromApp actions to the global batch
action store. A corresponding guardrail is added to the store to prevent
duplicate delete actions to be added to the basket.

Ref: SRX-GJ3E9Y

---------

Co-authored-by: Sven Urbanski <sven.urbanski@freiheit.com>