feat: simplify undeployment (#2668)
This change brings simplification of the application undeployment
process. It eliminates the concept of a multi-stage undeploy process
(prepare -> execute) that was dependent on special "undeploy version"
releases. The system is moving to a single, direct undeploy action that
immediately removes the application's deployment state, though removal
from individual environments also still is kept possible.