fix: sync basket (#2669)
When a user opens the "Delete Environments" dialog for an application,
any environments they have already selected for deletion (which are
stored as pending batch actions) will now be checked by default. This
improves the user experience by making the dialog's initial state
consistent with the current batch action list. Just keeping the state
from the last open dialog would not be sufficient as batch actions can
be _removed_ from the basket when the dialog is closed.

Ref: SRX-GJ3E9Y