fix(frontend): refactor env var parsing (#2677)
This changes how we parse environment variables in the frontend-service.
    
We now avoid envconfig.Process, because it makes it harder to find
environment variables by name (in the code).
    
The old way of parsing was kept for now - just to showcase that the test
behaves the same way with the old and new parsing.
    
The main function uses the new parser already.
    
Ref: SRX-4V1TTT