fix: update git sync status when deleting apps
Kuberpult now deletes the git sync status from the database,
whenever we delete an env from an app (or an entire app or entire env).

If you see an issue with the git sync status
(e.g. the status never turns green),
thne most likely there are old sync status in the DB
of apps or environments that were deleted.
In this case, you can execute this query:

TRUNCATE git_sync_status ;

The table git_sync_status only stores the current state,
so it will be refilled the next time kuberpult gets a request.

Ref: SRX-G4C6ZY